bullet

+++ basic-websocket-gameserver/src/main/java/com/bknote71/basicwebsocketgameserver/protocol/info/ProjectileInfo.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class ProjectileInfo {
-    private float speed;
+    private double speed;
     private double range;
 }
